backend/users/serializers.py
from rest_framework.serializers import ModelSerializer
from .models import User

class UserSerializer(ModelSerializer):
    class Meta:
        model = User
        fields = ['first_name', 'last_name', 'email', 'password', 'account_balance']
        extra_kwargs = {
            'password': {'write_only': True}
        }